requirements.txt
opencv-python~=4.5.2.52
numpy~=1.20.3
requests~=2.25.1
Pillow~=8.2.0
